actu

+++ database/migrations/2023_06_30_005125_documentacion_usuario.php
@@ -9,25 +9,38 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
-    {
-        Schema::disableForeignKeyConstraints();
-        Schema::create('documentacion_publicacion', function (Blueprint $table) {
+    public function up(): void
+{
+    Schema::disableForeignKeyConstraints();
+
+    if (!Schema::hasTable('documentacion_usuario')) {
+        Schema::create('documentacion_usuario', function (Blueprint $table) {
             $table->unsignedBigInteger('documentacion_id');
-            $table->unsignedBigInteger('publicacion_id');
+            $table->unsignedBigInteger('users_id');
             $table->timestamps();
         
-            $table->primary(['documentacion_id', 'publicacion_id']);
+            $table->primary(['documentacion_id', 'users_id']);
         
             $table->foreign('documentacion_id')->references('id')->on('documentacion')->onDelete('cascade');
-            $table->foreign('publicacion_id')->references('id')->on('publicacion')->onDelete('cascade');
+            $table->foreign('users_id')->references('id')->on('users')->onDelete('cascade');
         });
-        Schema::enableForeignKeyConstraints();
-
     }
 
-    public function down()
+    Schema::enableForeignKeyConstraints();
+}
+
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
     {
-        Schema::dropIfExists('documentacion_publicacion');
+        /// Elimina las referencias foráneas antes de eliminar la tabla
+    Schema::table('documentacion_usuario', function (Blueprint $table) {
+        $table->dropForeign(['users_id']);
+    });
+
+    // Elimina la tabla
+    Schema::dropIfExists('documentacion_usuario');
     }
 };
